requirements/requirements.txt
appdirs==1.4.4
asgiref==3.3.0
astroid==2.4.2
certifi==2020.6.20
distlib==0.3.1
Django==3.0.5
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.1
djongo==1.3.3
filelock==3.0.12
gunicorn==20.0.4
isort==5.5.2
lazy-object-proxy==1.4.3
Markdown==3.3.2
mccabe==0.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pymongo==3.11.0
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
virtualenv==20.0.35
wrapt==1.12.1
